Move IntTypeFormat to interp namespace.
