+++ src/intcomp/AbbreviationCodegen.cpp
@@ -24,6 +24,7 @@ namespace wasm {
 
 using namespace decode;
 using namespace filt;
+using namespace interp;
 
 namespace intcomp {
